SOA/Personne 2/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
psycopg2-binary==2.9.9
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.0
python-decouple==3.8
django-redis==5.4.0
drf-yasg==1.21.7
Faker==20.1.0
numpy==1.26.2